requirements.txt
pandas
numpy
scipy
matplotlib
PyYAML
ib-insync
ib-async